requirements.txt
beautifulsoup4==4.12.3
chromadb==0.6.3
fastapi==0.115.7
html2text==2024.2.26
ipfshttpclient==0.8.0
langchain-community==0.3.15
langchain-core==0.3.31
langchain-openai==0.3.1
numpy==1.26.4
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
selenium==4.29.0
sentence-transformers==3.4.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
uvicorn==0.34.0
webdriver-manager==4.0.2
pywin32==308
python-multipart==0.0.20 